let edit delete check token true or false

+++ app/api/createPost/route.ts
@@ -67,6 +67,15 @@ export async function GET(req: Request) {
 export async function DELETE(req: Request) {
     await connectDB()
 
+    const cookiesStore = await cookies()
+    const token = cookiesStore.get("token")?.value
+
+    if (!token) {
+        return Response.json({ message: "Unauthorized"}, { status: 401 })
+    }
+
+    // const decode = jwt.verify(token, "MY_SECRET_KEY") as JwtgetPost
+
     const body = await req.json()
 
     const deletePost = await Post.findByIdAndDelete(body._id)
@@ -76,6 +85,13 @@ export async function DELETE(req: Request) {
 
 export async function PUT(req: Request) {
     await connectDB()
+
+    const cookiesStore = await cookies()
+    const token = cookiesStore.get("token")?.value
+
+    if (!token) {
+        return Response.json({ message: "Unauthorized"}, { status: 401 })
+    }
 
     const body = await req.json()
